Improve behavior of __me().
Function now throws an exception if you attempt to retrieve a value that
cannot be echoed (i.e., one that cannot be converted into a string).

I would prefer to do some kind of duck-typing here (i.e., if it cannot
be echoed, throw an exception) but I'm not sure how to suppress Notices
into Exceptions, so for now it just checks whether the variable is in a
hard-coded subset of things I know are echoable (specifically, strings
and numeric values).